improvement(LegDiagramPreview): Add localization support.
