+++ packages/itinerary-body/src/__unpublished__/messages.d.ts
@@ -1,0 +1,2 @@
+// Generic module definition for all YAML file imports.
+declare module "*.yml";
